fix achivmentC

+++ src/achievement-p/achievement-p.controller.ts
@@ -5,7 +5,7 @@ import {
   Body,
   Patch,
   Param,
-  Delete,
+  Delete,ParseIntPipe,BadRequestException,NotFoundException
 } from '@nestjs/common';
 import { AchievementPService } from './achievement-p.service';
 import { CreateAchievementPDto } from './dto/create-achievement-p.dto';
@@ -15,31 +15,49 @@ import { UpdateAchievementPDto } from './dto/update-achievement-p.dto';
 export class AchievementPController {
   constructor(private readonly achievementPService: AchievementPService) {}
 
+  // 하나씩 추가
   @Post()
-  create(@Body() createAchievementPDto: CreateAchievementPDto) {
+  async create(@Body() createAchievementPDto: CreateAchievementPDto) {
     return this.achievementPService.create(createAchievementPDto);
   }
 
+// 전체조회
   @Get()
-  findAll() {
+  async findAll() {
     return this.achievementPService.findAll();
   }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.achievementPService.findOne(+id);
+  // Achievement타이틀 별 조회
+
+
+// 개별조회
+ @Get(':achievementPId')
+  async findOne(@Param('achievementPId') achievementPId: string) {
+  
+    return this.achievementPService.findOne(achievementPId);
   }
 
+  
+
+  // Achievement 타이틀과 일치하는 목록 추가
+
+  // 수행으로 수정
   @Patch(':id')
-  update(
-    @Param('id') id: string,
-    @Body() updateAchievementPDto: UpdateAchievementPDto,
+  async update(@Param('achievementId') achievementId: string,
+      @Body() updateAchievementPDto: UpdateAchievementPDto,
   ) {
-    return this.achievementPService.update(+id, updateAchievementPDto);
+    return this.achievementPService.update(achievementId,updateAchievementPDto)
+    
   }
+
+
+
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('achievementId') achievementId: string) {const id = Number(achievementId);
+    if (!id) {
+      throw new BadRequestException('achievementId 값이 없거나 타이에 맞지 않습니다');
+    }
     return this.achievementPService.remove(+id);
   }
 }
